moved create options functions to a separate file
